MODULES/requirements.txt
fastapi==0.98.0
MindsDB==23.5.3.2
mindsdb_sdk==1.0.2
MySQL-python==1.2.5
mysql_connector_repackaged==0.3.1
pandas==1.3.5
python-dotenv==1.0.0
PyYAML==6.0
